% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_survey.R
\docType{data}
\name{Class_survey}
\alias{Class_survey}
\title{Survey of students in a statistics class}
\format{A data frame with 27 rows, each of which is an individual student:
\itemize{
\item vehicle: Type of car the student drives
\item gender: Self-declared gender of the student
\item distance_campus: how far the student lives from the Lone Tree Campus of Co- conino Community College (miles)
\item ice_cream: favorite ice cream flavor
\item rent: How much the student pays in rent
\item major: Student's declared major
\item height: height of the student (inches)
\item winter: Student’s opinion of winter (Love it, Like it, Don’t like, No opinion)
}}
\usage{
data(Class_survey)
}
\description{
Survey results from two semesters of statistics classes at Coconino
Community College in the years 2018-2019.
}
\keyword{datasets}
