% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defects.R
\docType{data}
\name{Defects}
\alias{Defects}
\title{Defects in eyeglass manufacturing}
\format{A data.frame with one variable. Each row is
one pair of eyeglasses manufactured by Eyeglass-omatic}
\source{
Personal communication from John Matic,
a consultant to Eyeglass-omatic in Australia. (The company’s name is fictitious, but the data is from an actual company.)
}
\usage{
data(Defects)
}
\description{
Types of defects that an Eyeglass company sees in the lenses they
make into eyeglasses.
}
\keyword{datasets}
