% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eyeglasses.R, R/defects.R
\docType{data}
\name{Eyeglasses}
\alias{Eyeglasses}
\title{Processes used in eyeglass manufacturing}
\format{A data.frame with one variable. Each row is
one pair of eyeglasses manufactured by Eyeglass-omatic
\itemize{
\item \code{activity} the operation performed on the eyeglasses. For each
pair of glasses, the company performeds a single operation:
\code{Assemble}, \code{Grind}, put the \code{frames} together. \code{received} means
the eyeglasses were already totally finished. Guess what \code{Unknown} means!
}}
\source{
Personal communication from John Matic,
a consultant to Eyeglass-omatic in Australia. (The company’s name is fictitious, but the data is from an actual company.)

Personal communication from John Matic,
a consultant to Eyeglass-omatic in Australia. (The company’s name is fictitious, but the data is from an actual company.)
}
\usage{
data(Eyeglasses)

data(Eyeglasses)
}
\description{
Types of defects that an Eyeglass company sees in the lenses they
make into eyeglasses.
}
\keyword{datasets}
