% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Firearm.R
\docType{data}
\name{Firearm}
\alias{Firearm}
\title{Deaths from firearms}
\format{A data.frame wherein each row is year}
\source{
Deaths from firearms. (2013, September 26).
}
\usage{
data(Firearm)
}
\description{
The number of deaths caused by Firearms in Australia from 1983 to 1997,
expressed as a rate per 100,000 population.
}
\references{
Australian Institute of Criminology, 1999.The data was contributed by Rex Boggs, Glenmore State High School, Rockhampton, Queens- land, Australia.
}
\keyword{datasets}
