% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/House.R
\docType{data}
\name{House}
\alias{House}
\title{Capital and rental values of Auckland, New Zealand properties.}
\format{A data.frame wherein each row is one house
\itemize{
\item \code{Capital}: Selling price of house in Australian dollar (AUD)
\item \code{rental}: rental price of a house in Australian dollar (AUD)
}}
\source{
Source Capital and rental values of Auckland properties.
(2013, September 26). Retrieved from http://www.statsci.org/data/oz/rentcap.html
}
\usage{
data(House)
}
\description{
The data show the capital value and annual rental value of
domestic properties in Auckland in 1991.
}
\references{
Lee, A. (1994) Data Analysis:
An introduction based on R. Auckland: Department of Statistics,
University of Auckland.
Data courtesy of Sage Consultants Ltd.
}
\keyword{datasets}
