% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metacarpal.R
\docType{data}
\name{Metacarpal}
\alias{Metacarpal}
\title{Estimates of living stature from skeletal remains}
\format{A data.frame wherein each row is one skeleton
\itemize{
\item \code{length}: length of Metacarpal I bone in cm
\item \code{height}: stature of skeleton in cm
}}
\source{
Prediction of Height from Metacarpal Bone Length. (n.d.). Retrieved July 9, 2019, from http://www.statsci.org/data/general/stature.html
}
\usage{
data(Metacarpal)
}
\description{
When anthropologists analyze human skeletal remains, an important
piece of information is living stature. Since skeletons are
commonly based on statistical methods that utilize measurements on small bones.
The following data was presented in a paper in the \emph{American Journal of Physical Anthropology}
to validate one such  method.
}
\references{
Musgrave, J., and Harneja, N. (1978). The estimation of adult stature
from metacarpal bone length. Amer. J. Phys. Anthropology 48, 113- 120.
Devore, J., and Peck, R. (1986). Statistics.
The Exploration and Analysis of Data. West Publishing, St Paul, Minnesota.
}
\keyword{datasets}
