% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuition.R
\docType{data}
\name{Tuition}
\alias{Tuition}
\title{Costs at 4-year colleges}
\format{A data.frame wherein each row is one 4-year college or university
\itemize{
\item \code{INSTITUTION}: Name of four year institution
\item \code{TYPE}: Type of four year institution, Public_4_year, Private_4_year, For_profit_4_year.
\item \code{STATE}: What state the institution resides
\item \code{ROOM_BOARD}: The cost of room and board at the institution ($)
\item \code{INSTATE_TUTION}: The cost of instate tuition ($)
\item \code{INSTATE_TOTAL}: The cost of room and board and instate tuition ($ per year)
\item \code{OUTOFSTATE_TUTION}: The cost of out of state tuition ($ per year)
\item \code{OUTOFSTATE_TOTAL}: The cost of room and board and out of state tuition ($ per year)
}}
\source{
"Tuition and Fees, 1998-99 Through 2018-19". \emph{Chronicle of Higher Education} (2018, December 31). Retrieved from https://www.chronicle.com/interactives/tuition-and-fees
}
\usage{
data(Eyeglasses)
}
\description{
Costs at 4-year colleges
}
\keyword{datasets}
