% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wages.R
\docType{data}
\name{Wages}
\alias{Wages}
\title{Weekly earnings of workers}
\format{A data.frame wherein each row is one occupation
\itemize{
\item \code{Occupation}: Occupations of workers.
\item \code{Numworkers}: The number of workers in each occupation (in thousands of workers)
\item \code{median_wage}: Median weekly wage ($)
\item \code{male_worker}: number of male workers (in thousands of workers)
\item \code{male_wage}: Median weekly wage of male workers ($)
\item \code{female_worker}: number of female workers (in thousands of workers)
\item \code{female_wage}: Median weekly wage of female workers ($)
}}
\source{
CPS News Releases. (n.d.). Retrieved July 8, 2019, from https://www. bls.gov/cps/#'
}
\usage{
data(Wages)
}
\description{
Median weekly earnings of full-time wage and salary workers by detailed occupation
and sex.
The Current Population Survey (CPS) is a monthly survey of households conducted by
the Bureau of Census for the Bureau of Labor Statistics. It provides a
comprehensive body of data on the labor force, employ- ment, unemployment,
persons not in the labor force, hours of work, earnings, and other demographic
and labor force characteristics.
}
\keyword{datasets}
